httpctx: cleanup
